# src/main/java/com/mmako/spring/service/models/covidstats/regions/Region.java
package com.mmako.spring.service.models.covidstats.regions;

public class Region {

    private String iso;
    private String name;

    public String getIso() {
        return iso;
    }

    public void setIso(String iso) {
        this.iso = iso;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }
}